+++ MatrixSDK/Crypto/Verification/Transactions/SAS/MXSASTransaction_Private.h
@@ -24,6 +24,11 @@ NS_ASSUME_NONNULL_BEGIN
 
 #pragma mark - Constants
 
+extern const struct MXSASAgreementProtocols {
+    __unsafe_unretained NSString *v1;
+    __unsafe_unretained NSString *v2;
+} MXSASAgreementProtocols;
+
 FOUNDATION_EXPORT NSArray<NSString*> *kKnownAgreementProtocols;
 FOUNDATION_EXPORT NSArray<NSString*> *kKnownHashes;
 FOUNDATION_EXPORT NSArray<NSString*> *kKnownMacs;
@@ -46,7 +51,9 @@ FOUNDATION_EXPORT NSArray<NSString*> *kKnownShortCodes;
 - (void)handleMac:(MXKeyVerificationMac*)macContent;
 
 - (NSString*)hashUsingAgreedHashMethod:(NSString*)string;
-- (NSData*)generateSasBytesWithTheirPublicKey:(NSString*)theirPublicKey requestingDevice:(MXDeviceInfo*)requestingDevice otherDevice:(MXDeviceInfo*)otherDevice;
+- (NSData*)generateSasBytesWithTheirPublicKey:(NSString*)theirPublicKey
+                             requestingDevice:(MXDeviceInfo*)requestingDevice sasPublicKey:(NSString*)sasPublicKey
+                                  otherDevice:(MXDeviceInfo*)otherDevice otherSasPublicKey:(NSString*)otherSasPublicKey;
 
 @end