+++ CHANGES.rst
@@ -1,14 +1,31 @@
-Changes in Matrix iOS SDK in 0.16.3 (2020-xx-xx)
+Changes in Matrix iOS SDK in 0.16.5 (2020-05-xx)
 ================================================
 
 Improvements:
+ * MXSession: Update account data as soon as the endpoint returns.
+ * MXSecretStorage: Add this class to support SSSS ([MSC1946(]https://github.com/matrix-org/matrix-doc/pull/1946).
+ * SAS verification: Support new key agreement.
+
+Changes in Matrix iOS SDK in 0.16.4 (2020-05-07)
+================================================
+
+Improvements:
+ * Minimal version for iOS is now 9.0.
  * Pod: Update AFNetworking version (#793).
+ * Pod: Update Realm and OHTTPStubs.
+
+Changes in Matrix iOS SDK in 0.16.3 (2020-05-07)
+================================================
+
+Improvements:
+ * MXCrypto: Allow to verify a device again to request private keys again from it.
+ * Secrets: Validate received private keys for cross-signing and key backup before using them (vector-im/riot-ios/issues/3201).
 
 Changes in Matrix iOS SDK in 0.16.2 (2020-04-30)
 ================================================
 
 Improvements:
-* Cross-signing: Make key gossip requests when the other device sent m.key.verification.done (vector-im/riot-ios/issues/3163).
+ * Cross-signing: Make key gossip requests when the other device sent m.key.verification.done (vector-im/riot-ios/issues/3163).
 
 Bug fix:
  * MXEventTimeline: Fix crash in paginate:.